+++ crates/spaces/src/builtins/checkout.rs
@@ -310,24 +310,6 @@ pub const FUNCTIONS: &[Function] = &[
         example: Some(ADD_CARGO_BIN_EXAMPLE)
     },
     Function {
-        name: "add_capsule",
-        description: "Adds a capsule to the workspace during checkout.",
-        return_type: "None",
-        args: &[
-            get_rule_argument(),
-            Arg{
-                name : "capsule",
-                description: "dict with",
-                dict: &[
-                    ("scripts", "`spaces` scripts that define the capsule"),
-                    ("prefix", "Workspace prefix to install the capsule (`sysroot` for build deps and `build/install` for runtime deps)"),
-                    ("globs", "globs to apply to the capsule for installing at `prefix`"),
-                ]
-            }
-        ],
-        example: None,
-    },
-    Function {
         name: "add_asset",
         description: ADD_ASSET_DESCRIPTION,
         return_type: "None",
@@ -755,34 +737,6 @@ pub fn globals(builder: &mut GlobalsBuilder) {
             executor::Task::AddAsset(add_asset),
         ))
         .context(format_context!("Failed to insert task {rule_name}"))?;
-        Ok(NoneType)
-    }
-
-    fn add_capsule(
-        #[starlark(require = named)] rule: starlark::values::Value,
-        #[starlark(require = named)] capsule: starlark::values::Value,
-    ) -> anyhow::Result<NoneType> {
-        let rule: rules::Rule = serde_json::from_value(rule.to_json_value()?)
-            .context(format_context!("bad options for capsule rule"))?;
-
-        let capsule: executor::capsule::Capsule = serde_json::from_value(capsule.to_json_value()?)
-            .context(format_context!("Failed to parse capsule arguments"))?;
-
-        if let Some(globs) = capsule.globs.as_ref(){
-            changes::glob::validate(globs).context(format_context!("Failed to validate globs"))?;
-        }
-
-        let rule_name = rule.name.clone();
-
-        rules::insert_task(rules::Task::new(
-            rule,
-            rules::Phase::Checkout,
-            executor::Task::Capsule(capsule),
-        ))
-        .context(format_context!("Failed to insert task {rule_name}"))?;
-
-        // insert a Run rule also
-
         Ok(NoneType)
     }